// src/web-sockets/socket-events.enum.ts
export enum WsChatEventsEnum {
  SEND_MESSAGE = "send_message",
  RECEIVE_MESSAGE = "receive_message",
  EDIT_MESSAGE = "edit_message",
  DELETE_MESSAGE = "delete_message",
  RECEIVE_EDITED_MESSAGE = "receive_edited_message",
  MESSAGE_DELETED = "message_deleted",
  SEND_BROADCAST_MESSAGE = "send_broadcast_message",
}